[sw/unittests] Move dif unittests to sw/lib/dif.
This commit is part of lowrisc/opentitan#6107.

DIF unittests now live next to their test subjects in the `sw/lib/dif`
folder.

Added a dif test suite to be able to run DIF unittests as follows:

`meson test --suite=dif -C build-out`

Mock MMIO and CSR modules are now located under `sw/lib/base/testing` to
solve DIF dependency issues with the `sw/lib/testing` folder.
