+++ sw/device/lib/base/testing/mock_csr.cc
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "sw/device/lib/testing/mock_csr.h"
+#include "sw/device/lib/base/testing/mock_csr.h"
 
 #include "sw/device/lib/base/csr.h"
