+++ sw/device/lib/dif/dif_aon_timer_unittest.cc
@@ -7,7 +7,7 @@
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/bitfield.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "aon_timer_regs.h"  // Generated.
