+++ sw/device/lib/dif/dif_csrng_unittest.cc
@@ -9,7 +9,7 @@
 
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "csrng_regs.h"  // Generated
