+++ sw/device/lib/dif/dif_gpio_unittest.cc
@@ -8,7 +8,7 @@
 
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "gpio_regs.h"  // Generated
