+++ sw/device/lib/dif/dif_i2c_unittest.cc
@@ -10,7 +10,7 @@
 
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "i2c_regs.h"  // Generated.
