+++ sw/device/lib/dif/dif_lc_ctrl_unittest.cc
@@ -11,7 +11,7 @@
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/bitfield.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "lc_ctrl_regs.h"  // Generated.
