+++ sw/device/lib/dif/dif_otp_ctrl_unittest.cc
@@ -10,7 +10,7 @@
 
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "otp_ctrl_regs.h"  // Generated.
