+++ sw/device/lib/dif/dif_plic_unittest.cc
@@ -8,7 +8,7 @@
 
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "rv_plic_regs.h"  // Generated.
