+++ sw/device/lib/dif/dif_spi_device_unittest.cc
@@ -8,7 +8,7 @@
 
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 
 #include "spi_device_regs.h"  // Generated.
