+++ sw/device/tests/dif/meson.build
@@ -2,294 +2,6 @@
 # Licensed under the Apache License, Version 2.0, see LICENSE for details.
 # SPDX-License-Identifier: Apache-2.0
 
-test('dif_spi_device_unittest', executable(
-  'dif_spi_device_unittest',
-  sources: [
-    hw_ip_spi_device_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_spi_device.c',
-    'dif_spi_device_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_uart_unittest', executable(
-  'dif_uart_unittest',
-  sources: [
-    'dif_uart_unittest.cc',
-    meson.source_root() / 'sw/device/lib/dif/dif_uart.c',
-    hw_ip_uart_reg_h,
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_plic_unittest', executable(
-  'dif_plic_unittest',
-  sources: [
-    hw_top_earlgrey_rv_plic_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_plic.c',
-    'dif_plic_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_i2c_unittest', executable(
-  'dif_i2c_unittest',
-  sources: [
-    hw_ip_i2c_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_i2c.c',
-    'dif_i2c_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_gpio_unittest', executable(
-  'dif_gpio_unittest',
-  sources: [
-    hw_ip_gpio_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_gpio.c',
-    'dif_gpio_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_rv_timer_unittest', executable(
-  'dif_rv_timer_unittest',
-  sources: [
-    hw_ip_rv_timer_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_rv_timer.c',
-    'dif_rv_timer_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_otbn_unittest', executable(
-  'dif_otbn_unittest',
-  sources: [
-    'dif_otbn_unittest.cc',
-    meson.source_root() / 'sw/device/lib/dif/dif_otbn.c',
-    hw_ip_otbn_reg_h,
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_pwrmgr_unittest', executable(
-  'dif_pwrmgr_unittest',
-  sources: [
-    hw_ip_pwrmgr_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_pwrmgr.c',
-    'dif_pwrmgr_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_keymgr_unittest', executable(
-  'dif_keymgr_unittest',
-  sources: [
-    hw_ip_keymgr_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_keymgr.c',
-    'dif_keymgr_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_rstmgr_unittest', executable(
-  'dif_rstmgr_unittest',
-  sources: [
-    'dif_rstmgr_unittest.cc',
-    meson.source_root() / 'sw/device/lib/dif/dif_rstmgr.c',
-    hw_ip_rstmgr_reg_h,
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_alert_handler_unittest', executable(
-  'dif_alert_handler_unittest',
-  sources: [
-    'dif_alert_handler_unittest.cc',
-    meson.source_root() / 'sw/device/lib/dif/dif_alert_handler.c',
-    hw_ip_alert_handler_reg_h,
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_otp_ctrl_unittest', executable(
-  'dif_otp_ctrl_unittest',
-  sources: [
-    hw_ip_otp_ctrl_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_otp_ctrl.c',
-    'dif_otp_ctrl_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_lc_ctrl_unittest', executable(
-  'dif_lc_ctrl_unittest',
-  sources: [
-    hw_ip_lc_ctrl_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_lc_ctrl.c',
-    'dif_lc_ctrl_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_aes_unittest', executable(
-  'dif_aes_unittest',
-  sources: [
-    hw_ip_aes_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_aes.c',
-    'dif_aes_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_clkmgr_unittest', executable(
-  'dif_clkmgr_unittest',
-  sources: [
-    hw_ip_clkmgr_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_clkmgr.c',
-    'dif_clkmgr_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_csrng_unittest', executable(
-  'dif_csrng_unittest',
-  sources: [
-    hw_ip_csrng_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_csrng.c',
-    'dif_csrng_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_entropy_unittest', executable(
-  'dif_entropy_unittest',
-  sources: [
-    hw_ip_entropy_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_entropy.c',
-    'dif_entropy_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
-test('dif_aon_timer_unittest', executable(
-  'dif_aon_timer_unittest',
-  sources: [
-    hw_ip_aon_timer_reg_h,
-    meson.source_root() / 'sw/device/lib/dif/dif_aon_timer.c',
-    'dif_aon_timer_unittest.cc',
-  ],
-  dependencies: [
-    sw_vendor_gtest,
-    sw_lib_testing_mock_mmio,
-  ],
-  native: true,
-  c_args: ['-DMOCK_MMIO'],
-  cpp_args: ['-DMOCK_MMIO'],
-))
-
 dif_plic_smoketest_lib = declare_dependency(
   link_with: static_library(
     'dif_plic_smoketest_lib',
