+++ sw/device/tests/rom_ext/rom_ext_parser_unittest.cc
@@ -5,7 +5,7 @@
 #include "gtest/gtest.h"
 #include "sw/device/lib/base/bitfield.h"
 #include "sw/device/lib/base/mmio.h"
-#include "sw/device/lib/testing/mock_mmio.h"
+#include "sw/device/lib/base/testing/mock_mmio.h"
 #include "sw/device/silicon_creator/rom_exts/manifest.h"
 #include "sw/device/silicon_creator/rom_exts/rom_ext_manifest_parser.h"
